lab2/backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.1
Faker==40.12.0
gunicorn==25.3.0
packaging==26.0
psycopg[binary]==3.3.3
sqlparse==0.5.5
